Move haddock autolink to an extension.
[ci skip]
