+++ docs/.gitignore
@@ -1,1 +1,2 @@
 _build/*
+_extensions/*.pyc
